(Co)geodesic Flow

(Co)geodesic Flow

Methods for (co)geodesic flow. These are temporary, until Manifolds.jl provides the same functionality.

ManifoldHMC.cogeodesic_flow! — Function.

  cogeodesic_flow!(M, x, p, x₀, p₀, t)

Follow the geodesic with initial point $x_0 \in M$ and momentum $p_0 \in T_{x_0}^* M$ for time $t$, dragging $p_0$ along. The resulting point and cotangent vector are stored in $x$ and $p$, respectively.

source

ManifoldHMC.cogeodesic_flow — Method.

  cogeodesic_flow(M, x₀, p₀, t)

Follow the geodesic with initial point $x_0 \in M$ and momentum $p_0 \in T_{x_0}^* M$ for time $t$, dragging $p_0$ along.

source

ManifoldHMC.geodesic_flow! — Function.

  geodesic_flow!(M, x, v, x₀, v₀, t)

Follow the geodesic with initial point $x_0 \in M$ and velocity $v_0 \in T_{x_0} M$ for time $t$, dragging $v_0$ along. The resulting point and tangent vector are stored in $x$ and $v$, respectively.

source

ManifoldHMC.geodesic_flow — Method.

  geodesic_flow(M, x₀, v₀, t)

Follow the geodesic with initial point $x_0 \in M$ and velocity $v_0 \in T_{x_0} M$ for time $t$, dragging $v_0$ along.

source